+++ packages/moonstone/Panels/Router.js
@@ -37,7 +37,7 @@ const propTypes = {
 const Router = class extends React.Component {
 	static displayName = 'Router'
 
-	static propTypes = /** @lends moonstone/Panels.Router */ {
+	static propTypes = /** @lends moonstone/Panels.Router.prototype */ {
 		/**
 		 * List of views to render. Will be rendered as a flat array of views suitable for use in
 		 * Panels and not a hierarchy of views as the path implies.
@@ -187,13 +187,14 @@ const Router = class extends React.Component {
  */
 const Route = () => null;
 
-Route.propTypes = /** @lends moonstone/Popups.Route.prototype */ {
+Route.propTypes = {
 	/**
 	 * The component to render when the `path` for this Route matches the path of the
 	 * {@link moonstone/Panels.Routable} container.
 	 *
 	 * @type {String|Function}
 	 * @public
+	 * @memberof moonstone/Panels.Route.prototype
 	 */
 	component: React.PropTypes.oneOfType([
 		React.PropTypes.string,
@@ -205,6 +206,7 @@ Route.propTypes = /** @lends moonstone/Popups.Route.prototype */ {
 	 *
 	 * @type {String}
 	 * @public
+	 * @memberof moonstone/Panels.Route.prototype
 	 */
 	path: React.PropTypes.string.isRequired
 };