+++ packages/ui/ViewManager/TransitionGroup.js
@@ -1,8 +1,5 @@
-/**
+/*
  * Exports the {@link ui/ViewManager.TransitionGroup} component.
- *
- * @module ui/ViewManager/TransitionGroup
- * @private
  */
 
 // Using string refs from the source code of ReactTransitionGroup
@@ -84,12 +81,12 @@ const forwardOnWillTransition = forward('onWillTransition');
  * Currently somewhat specialized for the purposes of ViewManager.
  *
  * @class TransitionGroup
- * @memberof ui/ViewManager/TransitionGroup
+ * @memberof ui/ViewManager
  * @private
  */
 
 class TransitionGroup extends React.Component {
-	static propTypes = /** @lends ui/ViewManager/TransitionGroup.TransitionGroup.prototype */ {
+	static propTypes = /** @lends ui/ViewManager.TransitionGroup.prototype */ {
 		children: React.PropTypes.node.isRequired,
 
 		/**