+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,16 @@ The following is a curated list of changes in the Enact ui module, newest change
 ## Added
 
 - `onAppear`, `onEnter`, `onLeave`, `onStay`, `onTransition`, and `onWillTransition` events to ViewManager
+### Added
+
+- `none` to `scrimType` prop values in `ui/FloatingLayer`
+
+### Removed
+
+- `fit` prop from `ui/Transition` in favor of using `className`
+## [1.0.0-alpha.5] - 2016-12-16
+
+No changes.
 
 ## [1.0.0-alpha.4] - 2016-12-2